src/esm_eval/requirements.txt
# ESM Evaluation Pipeline Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
torch>=1.12.0
fair-esm>=2.0.0
matplotlib>=3.5.0
scipy>=1.7.0
numpy>=1.20.0

# Optional: for faster tokenization
sentencepiece>=0.1.96

